+++ phazer/CHANGELOG.md
@@ -1,3 +1,11 @@
 # v0.1.0
 
 Initial release: https://github.com/Coding-Badly/phazer
+
+# v0.1.1
+
+Initial release of the `simple` (synchronous) file creation helper (Phazer).
+
+# v0.1.2
+
+Initial release of the `tokio` (asynchronous) file creation helper (Phazer).